extract common reference controller methods into a module
